refactored webSocket Service
