+++ space-client/src/app/ws-client/space-web-socket.service.ts
